requirements.txt
# Core dependencies for Inebotten Discord selfbot

# Discord API
discord.py>=2.3.0

# HTTP client for API calls
aiohttp>=3.8.0

# Date/time utilities
python-dateutil>=2.8.0

# Safe expression evaluation
simpleeval>=0.9.0

# Google Calendar integration (optional)
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
